+++ docs/conf.py
@@ -26,9 +26,9 @@ copyright = '2018, The Laboratory of Experimental Biophysics, EPFL, Lausanne, Sw
 author = 'Baptiste Ottino, Kyle M. Douglass'
 
 # The short X.Y version
-version = '0.0.2'
+version = '0.1.0-SNAPSHOT'
 # The full version, including alpha/beta/rc tags
-release = '0.0.2'
+release = '0.1.0-SNAPSHOT'
 
 
 # -- General configuration ---------------------------------------------------